Cleanup PowerStatus interop
