+++ src/Common/src/Interop/Powrprof/Interop.GetTickCount.cs
@@ -1,0 +1,14 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Runtime.InteropServices;
+
+internal partial class Interop
+{
+    internal partial class Powrprof
+    {
+        [DllImport(Libraries.Powrprof, ExactSpelling = true)]
+        public static extern bool SetSuspendState(bool bHibernate, bool bForce, bool bWakeupEventsDisabled);
+    }
+}
