+++ src/System.Windows.Forms/src/System/Windows/Forms/Application.cs
@@ -1241,7 +1241,7 @@ namespace System.Windows.Forms
         ///  Returns true if the call succeeded, else false.
         /// </summary>
         public static bool SetSuspendState(PowerState state, bool force, bool disableWakeEvent)
-            => UnsafeNativeMethods.SetSuspendState(state == PowerState.Hibernate, force, disableWakeEvent);
+            => Powrprof.SetSuspendState(state == PowerState.Hibernate, force, disableWakeEvent);
 
         /// <summary>
         ///  Overload version of SetUnhandledExceptionMode that sets the UnhandledExceptionMode
