+++ src/System.Windows.Forms/src/System/Windows/Forms/PowerStatus.cs
@@ -2,11 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using static Interop;
+
 namespace System.Windows.Forms
 {
     public class PowerStatus
     {
-        private NativeMethods.SYSTEM_POWER_STATUS _systemPowerStatus;
+        private Kernel32.SYSTEM_POWER_STATUS _systemPowerStatus;
 
         internal PowerStatus()
         {
@@ -58,9 +60,6 @@ namespace System.Windows.Forms
             }
         }
 
-        private void UpdateSystemPowerStatus()
-        {
-            UnsafeNativeMethods.GetSystemPowerStatus(ref _systemPowerStatus);
-        }
+        private void UpdateSystemPowerStatus() => Kernel32.GetSystemPowerStatus(ref _systemPowerStatus);
     }
 }
